Empty commit to retrigger github actions
According to the red warning box of [1], a scheduled github workflow can
be auto disabled when it has a schedule but there's no repository
activity for 60 days, I think that's why our workflow is broken.

I manually disabled and then re-enabled the workflow (the workflow
itself does not say it's disabled). Use this new commit to try to
retrigger it to see if it works.

[1]: https://docs.github.com/en/actions/using-workflows/disabling-and-enabling-a-workflow